% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblids-package.R
\docType{package}
\name{biblids-package}
\alias{biblids}
\alias{biblids-package}
\title{biblids: Working with bibliometric identifiers}
\description{
Provides S3 classes, methods and predicates to work with bibliometric identifiers such as DOIs. Optionally (with `Suggests` dependencies) includes presentation methods, shiny input modules and some API wrappers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/biblids}
  \item \url{https://github.com/subugoe/biblids}
  \item Report bugs at \url{https://github.com/subugoe/biblids/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{held@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
  \item German Research Foundation [sponsor]
}

}
\keyword{internal}
