% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_api}
\alias{doi_api}
\alias{get_doi_handles}
\alias{resolve_doi}
\alias{is_doi_found}
\title{Use the doi.org API}
\usage{
get_doi_handles(x, query = NULL, ...)

resolve_doi(x, ...)

is_doi_found(x, ...)
}
\arguments{
\item{x}{A vector as created by, or convertable to \code{\link[=doi]{doi()}}.}

\item{query}{A named list of \href{https://www.doi.org/factsheets/DOIProxy.html#query-parameters}{query parameters}.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\description{
Queries the
\href{https://www.doi.org/factsheets/DOIProxy.html#rest-api}{DOI resolution proxy server REST API}.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{get_doi_handles}: Query the handles endpoint.
For possible queries, see \href{https://www.doi.org/factsheets/DOIProxy.html#rest-api}{doi.org}.
}
\itemize{
\item \code{resolve_doi}: Get the resolved URL for a DOI.
Returns \code{NA} if there is no URL value (rare).
}
\itemize{
\item \code{is_doi_found}: Test whether DOI handle can be found on doi.org.
}}}

\section{Warning}{

If you are using this in your own package, or create a lot of traffic,
please set your own \code{\link[httr:user_agent]{httr::user_agent()}}.
}

\section{Other APIs}{

This client only queries the doi.org API.
The doi.org API only includes information on DOI resolution,
not other metadata.

For additional metadata, you must access \emph{registration} agency APIs, such as
crossref, wrapped in \href{https://github.com/ropensci/rcrossref}{rcrossref}.
}

\examples{
# from https://www.doi.org/factsheets/DOIProxy.html
get_doi_handles(
  x = c("10.1000/1", "10.1038/nphys1170"),
  query = list(type = "URL")
)
resolve_doi(c("10.1038/nphys1170", "10.1000/1"))
is_doi_found(c("10.1000/I_DO_NOT_EXIST", "10.1000/1"))
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi}()},
\code{\link{regex_doi}()}
}
\concept{doi}
