% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_examples}
\alias{doi_examples}
\title{Some example dois}
\usage{
doi_examples()
}
\description{
Some example dois
}
\examples{
doi_examples()
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_patterns}()},
\code{\link{doi}()}
}
\concept{doi}
