% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_ra}
\alias{doi_ra}
\alias{doi_ras}
\alias{get_doi_ra}
\alias{is_doi_from_ra}
\title{Get the DOI Registration Agency (RA)}
\usage{
doi_ras()

get_doi_ra(x, ...)

is_doi_from_ra(x, ra = names(doi_ras()), ...)
}
\arguments{
\item{x}{A vector created by, or convertable to \code{\link[=doi]{doi()}}.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}

\item{ra}{Character scalar, must be one of \code{names(doi_ras())}.}
}
\description{
Registrants do not register DOIs directly with the DOI Foundation,
but go through RAs.
}
\details{
Some of these RAs have their own APIs to access additional metadata.

Selected APIs with existing R wrappers include:
\itemize{
\item \href{http://crossref.org}{Crossref},
wrapped by the \href{https://docs.ropensci.org/rcrossref/}{rcrossref} R client,
\item \href{https://datacite.org}{Datacite},
wrapped by the \href{https://docs.ropensci.org/rdatacite/}{rdatacite} R client,
\item \href{https://op.europa.eu/}{OP},
(partially ?) wrapped bye the
\href{https://cran.r-project.org/web/packages/eurlex/index.html}{eurlex}
R client.
}

It's easy to confuse these more comprehensive wrappers
geared towards particular RAs or services
with those API wrappers included in biblids such as \code{\link[=get_doi_handles]{get_doi_handles()}},
because \emph{both types of API accept DOIs as inputs}.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{doi_ras}: All current DOI RAs, returned as a character string,
named with its short name.
Taken from \href{https://www.doi.org/RA_Coverage.html}{doi.org},
where you can also learn more about areas of coverage.
}
\itemize{
\item \code{get_doi_ra}: Get DOI RA using the doi.org
\href{https://www.doi.org/factsheets/DOIProxy.html#whichra}{Which RA?} service.
}
\itemize{
\item \code{is_doi_from_ra}: Test whether a DOI is registed by an RA
}}}

\section{Warning}{

If you are using this in your own package, or create a lot of traffic,
please set your own \code{\link[httr:user_agent]{httr::user_agent()}}.
}

\examples{
doi_ras()
get_doi_ra(c(
  "10.6220/joq.2012.19(1).01", # Airiti
  "10.13336/j.1003-6520.hve.20160308018", # CNKI
  "10.1037/0003-066X.59.1.29", # Crossref
  "10.3207/2959859860", # DataCite, from https://www.doi.org/demos.html
  "10.5240/7290-C8AD-12BA-4F93-3B07-7", # EIDR
  "10.7666/d.y351065", # ISTIC, from https://www.doi.org/demos.html
  "10.11467/isss2003.7.1_11", # JaLC, from https://www.doi.org/demos.html
  "10.3743/KOSIM.2011.28.2.117", # KISTI, from https://www.doi.or.kr
  "10.1430/8105", # mEDRA, from from https://www.doi.org/demos.html
  "10.2777/11739", # OP, from https://op.europa.eu
  "10.1000/6" # Public, the International DOI Foundation itself
))
is_doi_from_ra(x = c("10.1000/1", "10.1038/nphys1170"), ra = "Crossref")
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_api}},
\code{\link{doi_examples}()},
\code{\link{doi}()},
\code{\link{regex_doi}()}
}
\concept{doi}
