% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{is_doi}
\alias{is_doi}
\title{Validate a DOI}
\usage{
is_doi(x, type = c("cr-modern", "cr-jws", "regexpal"))
}
\arguments{
\item{x}{a character string giving a DOI}

\item{type}{a character string giving the type of validation to run.
Implemented as regular expressions (see source code).
Must be one these syntax specifications:
\itemize{
\item from \href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{crossref}
\itemize{
\item \code{"cr-modern"} currently used crossref DOIs.
\item \code{"cr-jws"} for DOIs created by John Wiley & Sons
}
\item \code{"regexpal"} from \href{https://www.regexpal.com/96948}{regexpal} (undocumented, not recommended)
}}
}
\description{
Validates a \href{http://doi.org}{Digital Object Identifier (DOI)}
}
\details{
To resolve DOIs to the associated metadata, you can use the \href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{rcrossref} R package.
}
\examples{
is_doi("10.5281/zenodo.3892950") # TRUE
is_doi("http://doi.org/10.5281/zenodo.3892951") # TRUE
is_doi("lorem ipsum") # FALSE

}
\concept{doi}
