% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{regex_doi}
\alias{regex_doi}
\alias{doi_patterns}
\alias{str_extract_doi}
\alias{str_extract_all_doi}
\title{Find DOIs with regular expressions}
\usage{
regex_doi(type = c("doi.org", "cr-modern"), ...)

doi_patterns(type = c("doi.org", "cr-modern"))

str_extract_doi(string)

str_extract_all_doi(string, type = "doi.org")
}
\arguments{
\item{type}{a character string giving the type of validation to run.
Implemented as regular expressions (see source code).
Must be one these syntax specifications:
\itemize{
\item \code{"doi.org"} from \href{https://www.doi.org/doi_handbook/2_Numbering.html#2.2}{doi.org}, via \href{https://stackoverflow.com/questions/27910/finding-a-doi-in-a-document-or-page}{stack-overflow} (recommended):
\item \code{"cr-modern"} from \href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{crossref}
See examples.
}}

\item{...}{
  Arguments passed on to \code{\link[stringr:modifiers]{stringr::regex}}
  \describe{
    \item{\code{pattern}}{Pattern to modify behaviour.}
    \item{\code{ignore_case}}{Should case differences be ignored in the match?}
    \item{\code{multiline}}{If \code{TRUE}, \code{$} and \code{^} match
the beginning and end of each line. If \code{FALSE}, the
default, only match the start and end of the input.}
    \item{\code{comments}}{If \code{TRUE}, white space and comments beginning with
\code{#} are ignored. Escape literal spaces with \code{\\ }.}
    \item{\code{dotall}}{If \code{TRUE}, \code{.} will also match line terminators.}
  }}

\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\description{
Find DOIs with regular expressions
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{doi_patterns}: Find DOI fields with regular expressions
}
\itemize{
\item \code{str_extract_doi}: Extract \emph{first} DOIs from character strings
}
\itemize{
\item \code{str_extract_all_doi}: Extract \emph{all} DOIs from character strings
}}}

\examples{
regex_doi("doi.org")
regex_doi("cr-modern")

str_extract_doi(string = c(
  "10.1594/PANGAEA.726855",  # nothing to do here
  "10.1594/GFZ.GEOFON.gfz2009kciu ",  # remove space
  " 10.1594/PANGAEA.667386", # remove space
  "doi:10.3207/2959859860", # remove DOI
  "http://dx.doi.org/10.3866/PKU.WHXB201112303", # parse URL
  "10.3972/water973.0145.db&", # remove forbidden symbol
  "foo bar" # no DOI here
))
str_extract_all_doi(string = c(
  "10.7666/d.y351065",  # nothing to do here
  "10.11467/isss2003.7.1_11 doi:10.7875/leading.author.2.e008",  # space separated
  "doi:10.1430/8105&doi:10.1392/BC1.0 ",  # separated by forbidden
  "10.5194/wes-2019-70\n10.5194/wes-5-819-202",  # separated by linebreak
  "quux" # no DOI here
))
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi}()}
}
\concept{doi}
