% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{resolve_doi}
\alias{resolve_doi}
\alias{is_doi_resolveable}
\alias{get_doi_handles}
\title{Resolve a DOI}
\usage{
is_doi_resolveable(x)

get_doi_handles(x, ...)
}
\arguments{
\item{x}{a vector as created by \code{\link[=doi]{doi()}}.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\description{
Resolve a DOI
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{is_doi_resolveable}: Test whether DOI can be resolved
}
\itemize{
\item \code{get_doi_handles}: Get DOI handles
}}}

\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi}()},
\code{\link{regex_doi}()}
}
\concept{doi}
