% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{str_extract_all_doi}
\alias{str_extract_all_doi}
\alias{doi_patterns}
\title{Extract DOIs from strings}
\usage{
str_extract_all_doi(string, type = "doi.org")

doi_patterns(type = c("doi.org", "cr-modern"))
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{type}{a character string giving the type of validation to run.
Implemented as regular expressions (see source code).
Must be one these syntax specifications:
\itemize{
\item \code{"doi.org"} from \href{https://www.doi.org/doi_handbook/2_Numbering.html#2.2}{doi.org}, via \href{https://stackoverflow.com/questions/27910/finding-a-doi-in-a-document-or-page}{stack-overflow} (recommended):
\item \code{"cr-modern"} from \href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{crossref}
See examples.
}}
}
\description{
Extracts all DOIs to be found in a character string.
Useful when there are \emph{several} DOIs in \emph{one} string.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{doi_patterns}: Regular expressions for DOIs
}}}

\examples{
doi_patterns("doi.org")
doi_patterns("cr-modern")

}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi}()}
}
\concept{doi}
