% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{view_doi_matches}
\alias{view_doi_matches}
\alias{view_doi_matches_perline}
\alias{view_doi_matchesOutput}
\alias{renderView_doi_matches}
\title{View HTML rendering of matched DOIs
htmlwidget to show how DOIs are matched in strings.}
\usage{
view_doi_matches(string)

view_doi_matches_perline(string)

view_doi_matchesOutput(outputId, width = "100\%", height = "auto")

renderView_doi_matches(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates an HTML widget (or a
\href{https://rstudio.github.io/promises/}{promise} of an HTML widget).}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
View HTML rendering of matched DOIs
htmlwidget to show how DOIs are matched in strings.
}
\details{
Highlighting DOIs matching the regular expression
can help users troubleshoot their input,
for example inside shiny apps.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{view_doi_matches_perline}: Accepts only a scalar string, but retains linebreaks.
Helpful for showing matches from input widgets.
}
\itemize{
\item \code{view_doi_matchesOutput}: Shiny output widget to show matched DOIs
}
\itemize{
\item \code{renderView_doi_matches}: Shiny render function to show matched DOIs
}}}

\examples{
view_doi_matches("lorem ipsum 10.1000/foo dolor ist 10.1000/bar")
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_api}},
\code{\link{doi_examples}()},
\code{\link{doi_ra}},
\code{\link{doi}()},
\code{\link{regex_doi}()}
}
\concept{doi}
