% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblids-package.R
\docType{package}
\name{biblids-package}
\alias{biblids}
\alias{biblids-package}
\title{biblids: Validating with bibliometric identifiers}
\description{
Provides lightweight validation and normalisation for bibliometric identifiers such as DOIs.
    Optionally (with `Suggests` dependencies) includes checkmate assertions, testthat expectations and shiny input modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/biblids}
  \item \url{https://github.com/subugoe/biblids}
  \item Report bugs at \url{https://github.com/subugoe/biblids/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
