% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi}
\alias{doi}
\alias{is_doi}
\alias{as_doi}
\alias{format.biblids_doi}
\alias{pillar_shaft.biblids_doi}
\alias{knit_print.biblids_doi}
\alias{is.na.biblids_doi}
\title{\href{http://doi.org}{Digital Object Identifiers (DOI)}}
\usage{
doi(prefix = character(), suffix = character())

is_doi(x)

as_doi(x, ...)

\method{format}{biblids_doi}(x, ..., protocol = FALSE)

\method{pillar_shaft}{biblids_doi}(x, ...)

\method{knit_print}{biblids_doi}(
  x,
  display = getOption("biblids.doi_display", default = "crossref"),
  inline = FALSE,
  ...
)

\method{is.na}{biblids_doi}(x, ...)
}
\arguments{
\item{prefix}{the naming authority.}

\item{suffix}{the unique string chosen by the registrant}

\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}

\item{protocol}{logical flag, whether to prepend \verb{doi:} handle protocol, as per the official \href{https://doi.org/doi_handbook/2_Numbering.html#2.6.1}{DOI Handbook}.}

\item{display}{character scaling, giving how to display a DOI.
Must be one of:
\itemize{
\item \code{"crossref"} (recommended) to apply their \href{https://www.crossref.org/education/metadata/persistent-identifiers/doi-display-guidelines/}{display guidelines}.
Appears identical to the \href{https://support.datacite.org/docs/datacite-doi-display-guidelines}{DataCite display guidelines}.
\item \code{"doi"}: to apply the DOI Foundation \href{https://www.doi.org/doi_handbook/2_Numbering.html#2.6}{presentation guidelines}.
}}

\item{inline}{logical flag, giving whether to render DOIs as a chunk output or inline R.
Usually set by knitr.}
}
\description{
S3 record class for DOIs.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{is_doi}: test for \code{biblids_doi} class
}
\itemize{
\item \code{as_doi}: cast DOIs from other forms
}}
\subsection{Methods extending \code{biblids_doi} class (by generic):}{
\itemize{
\item \code{format}: display a doi
}
\itemize{
\item \code{pillar_shaft}: pretty printing in \code{\link[tibble:tibble]{tibble::tibble()}}
}
\itemize{
\item \code{knit_print}: pretty printing in R markdown (when knitr is available):
DOIs are hyperlined to the doi.org resolution service.\if{html}{\out{<div class="r">}}\preformatted{library(knitr)
# defaults to crossref style (recommended)
doi_examples()[1:3]
}\if{html}{\out{</div>}}
\itemize{
\item \url{https://doi.org/10.1038/nphys1170}
\item \code{NA}
\item \url{https://doi.org/10.11467/isss2003.7.1_11}
}\if{html}{\out{<div class="r">}}\preformatted{# or use doi style
knitr::knit_print(doi_examples()[1:3], display = "doi")
}\if{html}{\out{</div>}}
\itemize{
\item \href{https://doi.org/10.1038/nphys1170}{\code{doi:10.1038/nphys1170}}
\item \code{NA}
\item \href{https://doi.org/10.11467/isss2003.7.1_11}{\code{doi:10.11467/isss2003.7.1_11}}
}

You can also include DOIs inline with \url{https://doi.org/10.1038/nphys1170}, \code{NA} and \url{https://doi.org/10.11467/isss2003.7.1_11}.
}
\itemize{
\item \code{is.na}: a DOI needs \code{prefix} \emph{and} \code{suffix}, otherwise it's \code{NA}
}}}

\examples{
doi(
  # example dois are from https://www.doi.org/demos.html
  prefix = c("10.1038", NA, "10.11467"),
  suffix = c("nphys1170", "foobar", "isss2003.7.1_11")
)
is_doi(doi_examples())
as_doi(c(
  # example DOIs are from https://www.doi.org/demos.html
  "10.1594/PANGAE.726855",
  "10.1594/GFZ.GEOFON.gfz2009kciu"
))
# there is extra pretty printing inside tibbles
tibble::tibble(doi_examples()[1:3])
# this can be constructed but will be NA
is.na(doi(prefix = "10.5194", suffix = NA))
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi_patterns}()}
}
\concept{doi}
