% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi}
\alias{doi}
\alias{as.character.biblids_doi}
\alias{format.biblids_doi}
\alias{pillar_shaft.biblids_doi}
\alias{knit_print.biblids_doi}
\title{\href{http://doi.org}{Digital Object Identifiers (DOI)}}
\usage{
doi(prefix, suffix)

\method{as.character}{biblids_doi}(x, ..., protocol = TRUE, subtle_slash = FALSE)

\method{format}{biblids_doi}(x, ...)

pillar_shaft.biblids_doi(x, ...)

\method{knit_print}{biblids_doi}(x, inline = FALSE, ...)
}
\arguments{
\item{prefix}{denoting a unique naming authority}

\item{suffix}{unique string chosen by the registrant}

\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}

\item{protocol}{logical flag, whether to prepend \verb{doi:} handle protocol, as per the official \href{https://doi.org/doi_handbook/2_Numbering.html#2.6.1}{DOI Handbook}.
Recommended unless protocol is already obvious from the context, such as in a column header.}

\item{subtle_slash}{logical flag, whether to use \link[pillar:style_subtle]{pillar::style_subtle} slash between prefix and suffix.
Only for internal use, inside tibble printing.}

\item{inline}{logical flag, giving whether to render DOIs as a chunk output or inline R.
Usually set by knitr.}
}
\description{
S3 class for DOIs.
DOIs are stored as records.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.character}: convert to character vector

\item \code{format}: pretty printing in the R console.
}}

\section{Optional knitr print method}{

Inside knitr, bare DOIs are printed with the protocol and a link to the doi.org resolution.\if{html}{\out{<div class="r">}}\preformatted{library(knitr)
doi_examples
}\if{html}{\out{</div>}}
\itemize{
\item \href{https://doi.org/10.1038/nphys1170}{\code{doi:10.1038/nphys1170}}
\item \href{https://doi.org/10.1002/0470841559.ch1}{\verb{doi:10.1002/0470841559.ch1}}
\item \href{https://doi.org/NA}{\code{NA}}
\item \href{https://doi.org/10.11467/isss2003.7.1_11}{\code{doi:10.11467/isss2003.7.1_11}}
}

You can also include DOIs inline \href{https://doi.org/10.1038/nphys1170}{\code{doi:10.1038/nphys1170}}, \href{https://doi.org/10.1002/0470841559.ch1}{\verb{doi:10.1002/0470841559.ch1}}, \href{https://doi.org/NA}{\code{NA}} and \href{https://doi.org/10.11467/isss2003.7.1_11}{\code{doi:10.11467/isss2003.7.1_11}}.
}

\examples{
doi_examples <- doi(
  # example dois are from https://www.doi.org/demos.html
  prefix = c("10.1038", "10.1002", NA, "10.11467"),
  suffix = c("nphys1170", "0470841559.ch1", "foobar", "isss2003.7.1_11")
)
# you can print bare dois
doi_examples
# there is extra pretty printing inside tibbles
tibble::tibble(doi_examples)
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}},
\code{\link{doi_patterns}()}
}
\concept{doi}
