% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi}
\alias{doi}
\alias{is_doi}
\alias{is_doi_ish}
\alias{as_doi}
\alias{format.biblids_doi}
\alias{pillar_shaft.biblids_doi}
\alias{knit_print.biblids_doi}
\alias{is.na.biblids_doi}
\title{\href{http://doi.org}{Digital Object Identifiers (DOI)}}
\usage{
doi(prefix = character(), suffix = character())

is_doi(x)

is_doi_ish(x)

as_doi(x, ...)

\method{format}{biblids_doi}(x, ..., protocol = FALSE)

\method{pillar_shaft}{biblids_doi}(x, ...)

\method{knit_print}{biblids_doi}(
  x,
  display = getOption("biblids.doi_display", default = "crossref"),
  inline = FALSE,
  ...
)

\method{is.na}{biblids_doi}(x, ...)
}
\arguments{
\item{prefix}{The naming authority.}

\item{suffix}{The unique string chosen by the registrant.}

\item{x}{A vector as created by, or convertable to \code{\link[=doi]{doi()}}.}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}

\item{protocol}{Logical flag, whether to prepend \verb{doi:} handle protocol,
as per the official
\href{https://doi.org/doi_handbook/2_Numbering.html#2.6.1}{DOI Handbook}.}

\item{display}{character scaling, giving how to display a DOI.
Must be one of:
\itemize{
\item \code{"crossref"} (recommended)
to apply their
\href{https://www.crossref.org/education/metadata/persistent-identifiers/doi-display-guidelines/}{display guidelines}.
Appears identical to the
\href{https://support.datacite.org/docs/datacite-doi-display-guidelines}{DataCite display guidelines}.
\item \code{"doi"}:
to apply the DOI Foundation
\href{https://www.doi.org/doi_handbook/2_Numbering.html#2.6}{presentation guidelines}.
}}

\item{inline}{Logical flag, giving whether to render DOIs as a chunk output or inline R.
Usually set by knitr.}
}
\description{
S3 record class for DOIs.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{is_doi}: Test for \code{biblids_doi} class
}
\itemize{
\item \code{is_doi_ish}: Test whether character vector could be converted to DOI
}
\itemize{
\item \code{as_doi}: Cast DOIs from other forms
}}
\subsection{Methods extending \code{biblids_doi} class (by generic):}{
\itemize{
\item \code{format}: Display a DOI
}
\itemize{
\item \code{pillar_shaft}: Print DOIs in \code{\link[tibble:tibble]{tibble::tibble()}}s
}
\itemize{
\item \code{knit_print}: Print DOIs in R markdown (when knitr is available):
DOIs are hyperlined to the doi.org resolution service.\if{html}{\out{<div class="r">}}\preformatted{library(knitr)
# defaults to crossref style (recommended)
doi_examples()[1:3]
}\if{html}{\out{</div>}}
\itemize{
\item \url{https://doi.org/10.1038/nphys1170}
\item \code{NA}
\item \url{https://doi.org/10.11467/isss2003.7.1_11}
}\if{html}{\out{<div class="r">}}\preformatted{# or use doi style
knitr::knit_print(doi_examples()[1:3], display = "doi")
}\if{html}{\out{</div>}}
\itemize{
\item \href{https://doi.org/10.1038/nphys1170}{\code{doi:10.1038/nphys1170}}
\item \code{NA}
\item \href{https://doi.org/10.11467/isss2003.7.1_11}{\code{doi:10.11467/isss2003.7.1_11}}
}

You can also include DOIs inline with \url{https://doi.org/10.1038/nphys1170}, \code{NA} and \url{https://doi.org/10.11467/isss2003.7.1_11}.
}
\itemize{
\item \code{is.na}: Detect if \code{prefix} and/\emph{or} \code{suffix} is missing
}}}

\examples{
doi(
  # example dois are from https://www.doi.org/demos.html
  prefix = c("10.1038", NA, "10.11467"),
  suffix = c("nphys1170", "foobar", "isss2003.7.1_11")
)
# DOIs are case insensitive and are compared as such
unique(as_doi(c("10.1000/foo", "10.1000/fOo")))
as_doi("10.1000/BAR") == as_doi("10.1000/bar")

is_doi(doi_examples())
is_doi_ish(c(
  "foo",  # not a DOI
  "10.1371/journal.pone.0042793",
  "https://doi.org/10.13039/100000001"
))
as_doi(c(
  # example DOIs are from https://www.doi.org/demos.html
  "10.1594/PANGAE.726855",
  " 10.1594/GFZ.GEOFON.gfz2009kciu ",  # leading/trailing spaces are removed
  "https://doi.org/10.1000/182",  # URL form is parsed
  "doi:10.1000/1", # DOI from is parsed
  "foo bar" # yields NA
))
\dontrun{
# there must be only one DOI per element
as_doi(c("10.1126/science.169.3946.635 10.6084/m9.figshare.97218"))
}
# there is extra pretty printing inside tibbles
tibble::tibble(doi_examples()[1:3])
# this can be constructed but will be NA
is.na(doi(prefix = "10.5194", suffix = NA))
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_api}},
\code{\link{doi_examples}()},
\code{\link{regex_doi}()}
}
\concept{doi}
