% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_examples}
\alias{doi_examples}
\title{Example DOIs}
\usage{
doi_examples(na.rm = TRUE)
}
\arguments{
\item{na.rm}{Logical scalar, whether to remove \code{NA}s.
Helpful for testing and documentation.}
}
\description{
Example DOIs
}
\examples{
doi_examples()
}
\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_api}},
\code{\link{doi_ra}},
\code{\link{doi}()},
\code{\link{regex_doi}()},
\code{\link{view_doi_matches}()}
}
\concept{doi}
