% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_patterns}
\alias{doi_patterns}
\alias{str_extract_all_doi}
\title{Choose a DOI validation pattern}
\usage{
doi_patterns(type = c("doi.org", "cr-modern"))

str_extract_all_doi(string, ...)
}
\arguments{
\item{type}{a character string giving the type of validation to run.
Implemented as regular expressions (see source code).
Must be one these syntax specifications:
\itemize{
\item \code{"doi.org"} from \href{https://www.doi.org/doi_handbook/2_Numbering.html#2.2}{doi.org}, via \href{https://stackoverflow.com/questions/27910/finding-a-doi-in-a-document-or-page}{stack-overflow} (recommended)
\item \code{"cr-modern"} from \href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{crossref}
}}

\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{...}{
  Arguments passed on to \code{\link[=doi_patterns]{doi_patterns}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a named character vector of \code{prefix} and \code{suffix}, each a raw string with a regular expression.
}
\description{
Choose a DOI validation pattern
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{str_extract_all_doi}: Extract all DOIs from a string
}}}

\seealso{
Other doi: 
\code{\link{doiEntry}},
\code{\link{doi_examples}()},
\code{\link{doi}()}
}
\concept{doi}
