% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doiEntry}
\alias{doiEntry}
\alias{doiEntryApp}
\alias{doiEntryUI}
\alias{doiEntryServer}
\title{Shiny Module for DOI input}
\usage{
doiEntryApp()

doiEntryUI(id)

doiEntryServer(id)
}
\arguments{
\item{id}{The id string to be namespaced (optional).}
}
\description{
Accept, validate and return DOIs in a shiny app.
}
\section{Functions}{
\itemize{
\item \code{doiEntryApp}: Test app

\item \code{doiEntryUI}: Module UI

\item \code{doiEntryServer}: Module server
}}

\seealso{
Other doi: 
\code{\link{is_doi}()}
}
\concept{doi}
